requirements.txt
west==1.3.0
robotframework==6.1
robotframework-retryfailed==0.2.0
psutil==5.9.3
pyyaml==6.0.*
telnetlib3==2.0.*
pyrenode3 @ git+https://github.com/antmicro/pyrenode3.git
ai-edge-litert
pyserial==3.5
